Save bandwidth, time and disk space by mirroring an existing clone
The foundation submodule in particular is huge to be recloning on every
deploy, but in any case it's wasteful not to reuse the object databases
from a previous deploy before fetching anything missing.

This commit tries to clone the repository from the last successful
deploy, and otherwise clones from GitHub as usual.